rename column_zone_map to zone_map_index
