+++ be/test/olap/CMakeLists.txt
@@ -58,11 +58,11 @@ ADD_BE_TEST(rowset/segment_v2/ordinal_page_index_test)
 ADD_BE_TEST(rowset/segment_v2/rle_page_test)
 ADD_BE_TEST(rowset/segment_v2/binary_dict_page_test)
 ADD_BE_TEST(rowset/segment_v2/segment_test)
-ADD_BE_TEST(rowset/segment_v2/column_zone_map_test)
 ADD_BE_TEST(rowset/segment_v2/row_ranges_test)
 ADD_BE_TEST(rowset/segment_v2/frame_of_reference_page_test)
 ADD_BE_TEST(rowset/segment_v2/block_bloom_filter_test)
 ADD_BE_TEST(rowset/segment_v2/bloom_filter_index_reader_writer_test)
+ADD_BE_TEST(rowset/segment_v2/zone_map_index_test)
 ADD_BE_TEST(tablet_meta_manager_test)
 ADD_BE_TEST(tablet_mgr_test)
 ADD_BE_TEST(rowset/rowset_meta_manager_test)
